remove functionaltestcontroller, rename functional-test-controller file
